fix a comments bug
