+++ src/commentsCounter.js
@@ -1,17 +1,7 @@
-const validateCommentsContainer = (commentsContainer) => {
-  if (!commentsContainer || typeof commentsContainer.querySelectorAll !== 'function') {
-    return false;
-  }
-  return true;
-};
-
-const commentsCounter = (commentsContainer) => {
-  if (!validateCommentsContainer(commentsContainer)) {
-    throw new Error('Invalid comments container');
-  }
-
-  const comments = commentsContainer.querySelectorAll('.comment').length;
-  return comments;
+const commentsCounter = () => {
+  const comments = document.querySelectorAll('.list-items').length;
+  const commentsNumber = document.getElementById('comments-header');
+  commentsNumber.innerText = `Comments (${comments})`;
 };
 
 export default commentsCounter;
