+++ src/popUp.js
@@ -1,10 +1,9 @@
 import { getMovieData, createComment } from './comments.js';
-import getTVShows from './getTvShow.js';
+import commentsCounter from './commentsCounter.js';
 
 // Pop up function
 const popUp = async (show) => {
   const popUpContainer = document.createElement('article');
-  show.comments = await (getTVShows(show.id));
   popUpContainer.id = 'pop-up';
   popUpContainer.innerHTML = `
     <div class="pop-container">
@@ -31,9 +30,10 @@ const popUp = async (show) => {
            </div>
          </div>
        </div>
-       <div id="pop-up-comments">
-        <h3>Comments (${show.comments.length})</h3>
-        <ul id="comments-list"></ul>
+       <div class="pop-up-comments">
+        <h3 id="comments-header"> Comments (0)</h3>
+        <ul id="comments-list">
+        </ul>
       </div>
       <div class="pop-up-form">
         <h3>Add a comment</h3>
@@ -45,6 +45,7 @@ const popUp = async (show) => {
       </div>
      </div>
    `;
+
   document.body.append(popUpContainer);
   show.comments = await (getMovieData(show.id));
 
@@ -52,7 +53,30 @@ const popUp = async (show) => {
   closeBtn.addEventListener('click', () => {
     popUpContainer.remove();
   });
-  const commentList = document.getElementById('comments-list');
+
+  const renderComments = async () => {
+    const commentList = document.getElementById('comments-list');
+    commentList.innerHTML = '';
+
+    // Fetch the updated comments and populate the comment list
+    const updatedComments = await getMovieData(show.id);
+    if (Array.isArray(updatedComments)) {
+      updatedComments.forEach((comment) => {
+        const itemList = document.createElement('li');
+        itemList.classList.add('list-items');
+        itemList.innerHTML = `
+      <div>
+        <span>${comment.creation_date} </span>
+        <span class="user-name">${comment.username}: </span>
+        <span>${comment.comment}</span>
+      </div>
+    `;
+        commentList.appendChild(itemList);
+      });
+    }
+    commentsCounter();
+  };
+  renderComments();
   const addCommentBtn = document.getElementById('comment-btn');
   const commentForm = document.getElementById('new-comment');
 
@@ -69,21 +93,7 @@ const popUp = async (show) => {
 
     // Clear existing comments before adding new ones
     commentForm.reset();
-    commentList.innerHTML = '';
-
-    // Fetch the updated comments and populate the comment list
-    const updatedComments = await getMovieData(show.id);
-    updatedComments.forEach((comment) => {
-      const itemList = document.createElement('li');
-      itemList.innerHTML = `
-      <div>
-        <span>${comment.creation_date} </span>
-        <span class="user-name">${comment.username}: </span>
-        <span>${comment.comment}</span>
-      </div>
-    `;
-      commentList.appendChild(itemList);
-    });
+    await renderComments();
   });
   // Add form submit event listener for validation
   commentForm.addEventListener('submit', (e) => {
